backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance>=0.2.54
haver==0.8.0
parallel-web
synthefy
python-dotenv==1.0.0
pandas==2.2.0
pandas-datareader==0.10.0
pydantic==2.5.3
aiohttp==3.9.1
openai>=1.0.0
markdown>=3.5.0
PyYAML==6.0.1
SQLAlchemy==2.0.23
psycopg[binary,pool]==3.1.18
typer[all]==0.12.3

